test(e2e): remove redundant legacy status e2e, port unique cases to harmony suite
