+++ e2e/harmony/status-harmony.e2e.ts
@@ -1,5 +1,6 @@
-import { IssuesClasses } from '@teambit/component-issues';
+import { IssuesClasses, MISSING_DEPS_SPACE } from '@teambit/component-issues';
 import { expect } from 'chai';
+import { statusFailureMsg } from '@teambit/legacy.constants';
 import { Helper } from '@teambit/legacy.e2e-helper';
 
 describe('status command on Harmony', function () {
@@ -123,6 +124,59 @@ describe('status command on Harmony', function () {
     });
     it('bit status should not throw', () => {
       expect(() => helper.command.status()).not.to.throw();
+    });
+  });
+  describe('when a component is created and added without its dependencies', () => {
+    let output;
+    before(() => {
+      helper.scopeHelper.reInitWorkspace();
+      helper.fs.createFile(
+        'comp1',
+        'comp1.js',
+        `require("./comp2");require("./comp3");require("./comp4");require("./comp5");require("./comp6");`
+      );
+      helper.fs.createFile('comp2', 'comp2.js', `require("./comp4");require("./comp5");`);
+      helper.fs.createFile('comp3', 'comp3.js', '');
+      helper.fs.createFile('comp4', 'comp4.js', '');
+      helper.fs.createFile('comp5', 'comp5.js', 'require("./comp6");');
+      helper.fs.createFile('comp6', 'comp6.js', '');
+      helper.command.addComponent('comp1', { i: 'comp1' });
+      helper.command.addComponent('comp5', { i: 'comp5' });
+    });
+    it('Should show missing dependencies', () => {
+      output = helper.command.runCmd('bit status');
+      expect(output).to.have.string('non-existing dependency files');
+      expect(output).to.have.string('comp1 ... issues found');
+      expect(output).to.have.string('comp1.js -> ./comp2, ./comp3, ./comp4, ./comp5, ./comp6');
+      expect(output).to.have.string('comp5.js -> ./comp6');
+      expect(output).to.have.string('comp5 ... issues found');
+      // Validate indentations is correct, nested deps should be indent 2 more
+      expect(output).to.have.string(`${MISSING_DEPS_SPACE}comp1.js`);
+      expect(output).to.have.string(`${MISSING_DEPS_SPACE}comp5.js`);
+    });
+  });
+  describe('dynamic import', () => {
+    before(() => {
+      helper.scopeHelper.reInitWorkspace();
+      helper.fixtures.createComponentBarFoo('const a = "./b"; import(a); require(a);');
+      helper.fixtures.addComponentBarFoo();
+      helper.command.compile();
+    });
+    it('status should not show the component as missing packages', () => {
+      const output = helper.command.runCmd('bit status');
+      expect(output).to.not.have.string(statusFailureMsg);
+    });
+  });
+  describe('import from the index file', () => {
+    before(() => {
+      helper.scopeHelper.reInitWorkspace();
+      helper.fs.outputFile('comp1/index.ts', `export { hello } from './foo';`);
+      helper.fs.outputFile('comp1/foo.ts', `export const hello = 'world';`);
+      helper.fs.outputFile('comp1/bar.ts', `import { hello } from '.';`);
+      helper.command.addComponent('comp1');
+    });
+    it('should show an ImportFromDirectory issue', () => {
+      helper.command.expectStatusToHaveIssue(IssuesClasses.ImportFromDirectory.name);
     });
   });
 });
